Restructure files based on PR feedback
